requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-multipart==0.0.12
PyJWT==2.10.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
email-validator==2.1.0
alembic==1.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0
httpx==0.27.2
